github-releases: findNewDigest to digest.ts
